Add staging option for webdriver test
