+++ .travis.yml
@@ -37,6 +37,13 @@ matrix:
         - DOCKER_INSTANCE=wptd-dev-${RANDOM}
         - INSTALL_TARGET=go_deps
         - TEST_TARGET=go_test
+    - language: go
+      env:
+        - DOCKER_FILE=Dockerfile.dev
+        - DOCKER_IMAGE=wptd-dev
+        - DOCKER_INSTANCE=wptd-dev-${RANDOM}
+        - INSTALL_TARGET=webapp_deploy_staging
+        - TEST_TARGET=go_webdriver_test_staging
     - language: python
       env:
         - DOCKER_FILE=Dockerfile.jenkins
@@ -46,6 +53,8 @@ matrix:
         - TEST_TARGET=jenkins_test
 
 before_install:
+  - openssl aes-256-cbc -K $encrypted_ddee49501f31_key -iv $encrypted_ddee49501f31_iv
+    -in client-secret.json.enc -out client-secret.json -d
   - docker build -t wptd-base -f Dockerfile.base .
   - docker build -t "${DOCKER_IMAGE}" -f "${DOCKER_FILE}" .
   - docker run -t -d --entrypoint /bin/bash -v /etc/group:/etc/group:ro -v /etc/passwd:/etc/passwd:ro -v "$(pwd)":/home/jenkins/wptdashboard -u $(id -u $USER):$(id -g $USER) --name "${DOCKER_INSTANCE}" "${DOCKER_IMAGE}"
@@ -55,7 +64,7 @@ install:
   - docker exec -u 0:0 "${DOCKER_INSTANCE}" chown -R $(id -u $USER):$(id -g $USER) /home/jenkins
 
 script:
-  - docker exec -t -u $(id -u $USER):$(id -g $USER) "${DOCKER_INSTANCE}" make "${TEST_TARGET}"
+  - if [ "${TEST_TARGET}" != "" ]; then docker exec -t -u $(id -u $USER):$(id -g $USER) "${DOCKER_INSTANCE}" make "${TEST_TARGET}"; fi
 
 after_script:
   - docker stop "${DOCKER_INSTANCE}"
