+++ git/git-deploy
@@ -8,19 +8,21 @@ source "${REPO_DIR}/util/path.sh"
 WPTD_PATH=${WPTD_PATH:-$(absdir "${REPO_DIR}")}
 
 usage() {
-  info "Usage: git-deploy [-p] [-h]";
+  info "Usage: git-deploy [-p] [-h] [-q]";
 }
 
-while getopts ':ph' flag; do
+PRODUCTION='false'
+QUIET='false'
+
+while getopts ':phq' flag; do
   case "${flag}" in
     p) PRODUCTION='true' ;;
+    q) QUIET='true' ;;
     h|*) usage && exit 0;;
   esac
 done
 
-BRANCH_NAME="$(git rev-parse --abbrev-ref HEAD)"
-USER="$(git remote -v get-url origin | sed -E 's#(https?:\/\/|git@)github.com(\/|:)##' | sed 's#/.*$##')"
-VERSION="${USER}-${BRANCH_NAME}"
+source ${REPO_DIR}/git/staging-version.sh
 PROMOTE="--no-promote"
 
 if [[ ${PRODUCTION} == 'true' ]]
@@ -28,19 +30,31 @@ then
   info "Producing production configuration..."
   if [[ "${USER}" != "w3c" ]]
   then
-    confirm "Are you sure you want to be deploying a non-w3c repo (${USER})?"
-    if [ "${?}" != "0" ]; then exit "${?}"; fi
+    if [[ "${QUIET}" != "true" ]]
+    then
+        confirm "Are you sure you want to be deploying a non-w3c repo (${USER})?"
+        if [ "${?}" != "0" ]; then exit "${?}"; fi
+    fi
   fi
   # Use SHA for prod-pushes.
   VERSION="$(git rev-parse --short HEAD)"
   PROMOTE="--promote"
 fi
 
-COMMAND="gcloud app deploy ${PROMOTE} --version=${VERSION} ${WPTD_PATH}/webapp"
+if [[ "${QUIET}" == "true" ]]
+then
+    QUIET_FLAG="-q"
+else
+    QUIET_FLAG=""
+fi
+COMMAND="gcloud app deploy ${PROMOTE} ${QUIET_FLAG} --version=${VERSION} ${WPTD_PATH}/webapp"
 
 info "Deploy command:\n${COMMAND}"
-confirm "Execute?"
-if [ "${?}" != "0" ]; then exit "${?}"; fi
+if [[ "${QUIET}" != "true" ]]
+then
+    confirm "Execute?"
+    if [ "${?}" != "0" ]; then exit "${?}"; fi
+fi
 
 info "Executing..."
 ${COMMAND}
